requirements.txt
dj-database-url==0.5.0
Django==2.0.6
djangorestframework==3.8.2
psycopg2==2.7.4
pytz==2018.4
